pyproject.toml
[tool.poetry]
name = "neural-style-transfer"
version = "0.1.0"
description = ""
authors = ["Philip Hierhager <philip.hierhager@googlemail.com>"]
readme = "README.md"
packages = [{include = "neural_style_transfer"}]

[tool.poetry.dependencies]
python = ">=3.9, <3.11"
pillow = "^10.0.0"
tensorflow-io-gcs-filesystem = "0.31.0"
tensorflow = "^2.13.0"

[tool.poetry.scripts]
train = "neural_style_transfer.train:train"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
